requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Agent Framework
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0

# Vector Store
chromadb>=0.4.22

# NLP Models
sentence-transformers>=2.3.1
vaderSentiment>=3.3.2
bertopic>=0.16.0
spacy>=3.7.0
pytextrank>=3.2.5
transformers>=4.35.0
torch>=2.0.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Validation & Config
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# API & File Handling
python-multipart>=0.0.9

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# Utilities
umap-learn>=0.5.5
hdbscan>=0.8.33
scikit-learn>=1.4.0

# Logging
colorlog>=6.8.2

# UI Framework
streamlit>=1.31.0

# Visualization Libraries
plotly>=5.18.0
altair>=5.2.0
wordcloud>=1.9.0
matplotlib>=3.8.0

# Export Libraries
reportlab>=4.0.0
fpdf2>=2.7.0
Pillow>=10.0.0
